auren/agents/neuros/requirements.txt
# NEUROS Cognitive Service Requirements
# Production dependencies for biometric-aware AI personality system

# Core LangGraph/LangChain - FIXED VERSIONS
langgraph==0.2.56
langchain==0.3.17
langchain-openai==0.3.28  # UPGRADED from 0.1.23
langchain-core==0.3.72
langgraph-checkpoint-redis==0.0.8  # Added official Redis checkpoint support

# Async support
asyncio>=3.4.3
asyncpg>=0.29.0
redis[hiredis]>=5.0.1

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Kafka
aiokafka>=0.10.0
confluent-kafka>=2.3.0

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
structlog>=23.2.0
orjson>=3.9.10

# Monitoring & OpenTelemetry (PRODUCTION COMPATIBLE VERSIONS)
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-kafka-python>=0.42b0
opentelemetry-exporter-prometheus>=0.57b0
opentelemetry-exporter-otlp>=1.21.0

# Health checks
httpx>=0.25.2

# YAML parsing
PyYAML>=6.0.1

# Environment management
python-dotenv>=1.0.0

# Error tracking (optional)
sentry-sdk>=1.39.1

# Testing (optional, remove for production)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0 

# ChromaDB for L3 semantic memory
chromadb==0.4.22 